Improvement/ Fixing some errors- improving some components- adding loading to group routes
