+++ src/app/api/followuser/route.tsx
@@ -9,10 +9,11 @@ export async function POST(request: NextRequest) {
 
   if (infoCookie?.value) {
     const infoCookieObject = JSON.parse(infoCookie.value);
+    const userId = infoCookieObject.userId;
     const token = infoCookieObject.token;
 
     const response = await fetch(
-      `http://wiishy-backend.ir/user-home`,
+      `http://wiishy-backend.ir/api/follow/${userId}/${req.userId}`,
       {
         method: "GET",
         headers: {
