+++ src/app/api/gptengine/route.tsx
@@ -1,7 +1,8 @@
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(req: NextRequest, res: NextResponse) {
-  const API_URL = 'https://api.openai.com/v1/engines/davinci/completions';
+  const API_URL = 'https://api.openai.com/v1/engines/davinci-codex/completions';
+  // const API_URL = 'https://api.openai.com/v1/engines/davinci-codex/completions';
   const { prompt } = await req.json();
 
   try {
@@ -10,6 +11,8 @@ export async function POST(req: NextRequest, res: NextResponse) {
       headers: {
         "Content-Type": "application/json",
         // 'Authorization': `Bearer ${process.env.OPENAI_API_KEY}`,
+        // My code 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
+        // Emad key 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
         Authorization: `Bearer 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`,
       },
       body: JSON.stringify({
