+++ src/app/profile/(surfing)/layout.tsx
@@ -18,7 +18,7 @@ export default function RootLayout({
   return (
     <>
     
-        <div className="pt-[50px]">
+        <div className="pt-[50px] h-full">
           <Header />
           {children}
           <Navbar />
